Extract upload/download tests into module for reuse
