+++ test/functional/backends/test_netssh_scp.rb
@@ -1,0 +1,17 @@
+require 'helper'
+require_relative 'netssh_transfer_tests'
+
+module SSHKit
+  module Backend
+    class TestNetsshScp < FunctionalTest
+      include NetsshTransferTests
+
+      def setup
+        super
+        SSHKit::Backend::Netssh.configure do |ssh|
+          ssh.transfer_method = :scp
+        end
+      end
+    end
+  end
+end
